Majority of Part 5
Created the classes required (card, subclasses of card, deckNode, deck, hand). Deck was implemented as a linked list, and Hand was implemented as a vector. Assignment operator and copy constructors have yet to be implemented. Play() method requires implementation of Orders before it can actually create any orders (though it does everything else that it should). I have been testing everything with main so the driver file has yet to be changed.